feat: architecture refactor using interfaces
