+++ src/main/java/com/uade/beappsint/controller/CustomerController.java
@@ -2,25 +2,16 @@ package com.uade.beappsint.controller;
 
 import com.uade.beappsint.dto.kyc.KycRequestDTO;
 import com.uade.beappsint.dto.kyc.KycResponseDTO;
-import com.uade.beappsint.service.CustomerService;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RequestMapping("/customer")
 @RestController
-public class CustomerController {
-    private final CustomerService customerService;
-
-    public CustomerController(CustomerService customerService) {
-        this.customerService = customerService;
-    }
-
+public interface CustomerController {
     @PostMapping("/kyc")
-    public ResponseEntity<KycResponseDTO> basicKyc(@RequestBody @Validated KycRequestDTO requestDTO) {
-        return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(customerService.basicKyc(requestDTO));
-    }
+    ResponseEntity<KycResponseDTO> basicKyc(@RequestBody @Validated KycRequestDTO requestDTO);
 }
