+++ src/main/java/com/uade/beappsint/service/ProductService.java
@@ -3,113 +3,31 @@ package com.uade.beappsint.service;
 import com.uade.beappsint.dto.ProductDTO;
 import com.uade.beappsint.entity.Customer;
 import com.uade.beappsint.entity.Product;
-import com.uade.beappsint.exception.BadRequestException;
-import com.uade.beappsint.repository.CustomerRepository;
-import com.uade.beappsint.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
-public class ProductService {
-    private final ProductRepository productRepository;
-    private final CustomerRepository customerRepository;
-    private final AuthService authService;
+public interface ProductService {
+    List<ProductDTO> getAllProducts();
 
+    ProductDTO getProductById(Long id);
 
-    public ProductService(ProductRepository productRepository, CustomerRepository customerRepository, AuthService authService) {
-        this.productRepository = productRepository;
-        this.customerRepository = customerRepository;
-        this.authService = authService;
-    }
+    ProductDTO createProduct(Product product);
 
-    public List<ProductDTO> getAllProducts() {
-        return productRepository.findAll()
-                .stream()
-                .map(Product::toDTO)
-                .collect(Collectors.toList());
-    }
+    ProductDTO updateProduct(Long id, Product productDetails);
 
-    public ProductDTO getProductById(Long id) {
-        Product product = productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
-        return product.toDTO();
-    }
+    void deleteProduct(Long id);
 
-    public ProductDTO createProduct(Product product) {
-        assertAdmin();
-        product.setCreatedBy(authService.getAuthenticatedCustomer());
-        Product savedProduct = productRepository.save(product);
-        return savedProduct.toDTO();
-    }
+    void viewProduct(Long productId);
 
-    public ProductDTO updateProduct(Long id, Product productDetails) {
-        Customer customer = assertAdmin();
-        isProductCreator(id, customer);
-        Product product = productRepository.findById(id).orElseThrow(() -> new RuntimeException("Product not found"));
-        product.setName(productDetails.getName());
-        product.setDescription(productDetails.getDescription());
-        product.setStock(productDetails.getStock());
-        product.setPrice(productDetails.getPrice());
-        product.setCategory(productDetails.getCategory());
-        product.setImageUrl(productDetails.getImageUrl());
-        Product updatedProduct = productRepository.save(product);
-        return updatedProduct.toDTO();
-    }
+    List<ProductDTO> getFeaturedProducts();
 
-    public void deleteProduct(Long id) {
-        assertAdmin();
-        isProductCreator(id, authService.getAuthenticatedCustomer());
-        productRepository.deleteById(id);
-    }
+    List<ProductDTO> getProductsByCategory(String category);
 
-    public void viewProduct(Long productId) {
-        Customer customer = authService.getAuthenticatedCustomer();
-        Product product = productRepository.findById(productId).orElseThrow(() -> new RuntimeException("Product not found"));
+    List<ProductDTO> getRecentlyViewedProducts();
 
-        product.setViews(product.getViews() + 1);
-        productRepository.save(product);
+    Customer assertAdmin();
 
-        customer.getRecentlyViewedProducts().add(product);
-        customerRepository.save(customer);
-    }
-
-    public List<ProductDTO> getFeaturedProducts() {
-        return productRepository.findTop10ByOrderByViewsDesc()
-                .stream()
-                .map(Product::toDTO)
-                .collect(Collectors.toList());
-    }
-
-    public List<ProductDTO> getProductsByCategory(String category) {
-        return productRepository.findByCategory(category)
-                .stream()
-                .map(Product::toDTO)
-                .collect(Collectors.toList());
-    }
-
-    public List<ProductDTO> getRecentlyViewedProducts() {
-        Customer customer = authService.getAuthenticatedCustomer();
-        return customer.getRecentlyViewedProducts()
-                .stream()
-                .map(Product::toDTO)
-                .collect(Collectors.toList());
-    }
-
-    public Customer assertAdmin() {
-        Customer customer = authService.getAuthenticatedCustomer();
-        if (!customer.isAdmin()) {
-            throw new RuntimeException("Access denied: only administrators can perform this action.");
-        }
-        return customer;
-    }
-
-    private void isProductCreator(Long productId, Customer customer) {
-        Product product = productRepository.findById(productId)
-                .orElseThrow(() -> new BadRequestException("Product not found"));
-
-        if (!product.getCreatedBy().getId().equals(customer.getId())) {
-            throw new RuntimeException("Access denied: you are not the creator of this product.");
-        }
-    }
+    void isProductCreator(Long productId, Customer customer);
 }
