+++ src/main/java/com/uade/beappsint/service/impl/AuthServiceImpl.java
@@ -1,0 +1,70 @@
+package com.uade.beappsint.service.impl;
+
+import com.uade.beappsint.dto.auth.*;
+import com.uade.beappsint.entity.Customer;
+import com.uade.beappsint.exception.BadRequestException;
+import com.uade.beappsint.exception.UserAlreadyExistsException;
+import com.uade.beappsint.repository.CustomerRepository;
+import com.uade.beappsint.service.AuthService;
+import com.uade.beappsint.service.JwtService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class AuthServiceImpl implements AuthService {
+    private final CustomerRepository customerRepository;
+    private final BCryptPasswordEncoder passwordEncoder;
+    private final JwtService jwtService;
+    private final AuthenticationManager authenticationManager;
+
+    public SignupResponseDTO signup(SignupRequestDTO request) {
+        if (customerRepository.existsByEmail(request.getEmail())) throw new UserAlreadyExistsException(request.getEmail());
+        if (request.getPassword() == null) throw new BadRequestException("Password is required.");
+
+        Customer newUser = Customer.builder()
+                .email(request.getEmail())
+                .password(passwordEncoder.encode(request.getPassword()))
+                .kycCompleted(false)
+                .isAdmin(false)
+                .build();
+
+        customerRepository.save(newUser);
+
+        return SignupResponseDTO.builder()
+                .message("The user was created.")
+                .build();
+    }
+
+    public LoginResponseDTO login(LoginRequestDTO request) {
+        authenticationManager.authenticate(
+                new UsernamePasswordAuthenticationToken(
+                        request.getEmail(),
+                        request.getPassword()
+                )
+        );
+
+        Customer customer = customerRepository.findByEmail(request.getEmail())
+                .orElseThrow();
+
+        String jwtToken = jwtService.generateToken(customer);
+        return LoginResponseDTO.builder()
+                .token(jwtToken)
+                .expiresIn(jwtService.getExpirationTime())
+                .build();
+    }
+
+    public Customer getAuthenticatedCustomer() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return (Customer) authentication.getPrincipal();
+    }
+
+    public CustomerInfoDTO getCustomerInfo() {
+        return getAuthenticatedCustomer().toDto();
+    }
+}
